XXE_LAB/backend/requirements.txt
flask
mysql-connector-python
lxml
